added comment section, truncate des on home page, image limit to 500kb

+++ app.py
@@ -2,10 +2,11 @@ from flask import Flask, render_template, redirect, request, flash
 from werkzeug.utils import secure_filename
 from flask_sqlalchemy import SQLAlchemy
 import sqlite3
-import datetime
+from datetime import datetime
 import smtplib
 import os
 from constants import *
+import csv_writer
 
 
 app = Flask(__name__)
@@ -14,14 +15,15 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite://///home/raghav/Desktop/kitab_a
 app.config['UPLOAD_FOLDER'] = "/home/raghav/Desktop/kitab_app/static/images"
 db = SQLAlchemy(app)
 
+
+
 s = smtplib.SMTP('smtp.gmail.com', 587)
 s.starttls()
-me = "helpinghandsforneedies@gmail.com"
-password = "bansal@123"
-s.login(me, password) 
+# s.login(me, password) 
 
 
 class Book(db.Model):
+
 	id = db.Column(db.Integer, primary_key = True)
 	username = db.Column(db.String(50), nullable=False)
 	email = db.Column(db.String(50), nullable=False)
@@ -29,17 +31,34 @@ class Book(db.Model):
 	category = db.Column(db.String(50), nullable=False)
 	description = db.Column(db.String(200), nullable=False)
 	phone = db.Column(db.Integer, nullable=False)
-	address = db.Column(db.String(100), nullable=False)
 	location = db.Column(db.String(50), nullable=False)
-	createdOn = db.Column(db.DateTime, default=datetime.datetime.utcnow)
+	createdOn = db.Column(db.DateTime, default=datetime.now())
 	image_path = db.Column(db.String(300), default="book.jpg")
+
+	def get_book_as_tuple(self, status):
+		return (self.username, self.email, self.title, self.category, self.phone, self.description, self.location, self.createdOn, datetime.now(), status)
 
 	def __repr__(self):
 		return "{} {}".format(self.email, self.title) 
 
+class Comments(db.Model):
+	__tablename__ = "Comments"
+
+	id = db.Column(db.Integer, primary_key = True)
+	pid = db.Column(db.Integer)
+	comment = db.Column(db.String(100), nullable=False)
+	createdOn = db.Column(db.DateTime, default=datetime.now())
+
+	def __repr__(self):
+		return "{} {}".format(self.pid, self.comment) 
 
 @app.route('/')
 def home():
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/", datetime.now())
+	)
 
 	def helper(s):
 		return "" if(s=="" or s=="-Select-" or s=="" or s==None) else s
@@ -50,32 +69,57 @@ def home():
 	
 	if(location=="" and category=="" and email==""):
 		books = Book.query.all()
-	elif(location=="" and email==""):
-		books = Book.query.filter_by(category=category)
-	elif(category=="" and email==""):
-		books = Book.query.filter_by(location=location)
-	elif(category=="" and location==""):
-		books = Book.query.filter_by(email=email)
-	elif(email==""):
-		books = Book.query.filter_by(location=location, category = category)
-	elif(location==""):
-		books = Book.query.filter_by(email=email, category = category)
-	elif(category==""):
-		books = Book.query.filter_by(location=location, email = email)
 	else:
-		books = Book.query.filter_by(location=location, email=email, category = category)
+
+		csv_writer.write(
+			"/home/raghav/Desktop/kitab_app/user_data/searchs.csv",
+			(location, email, category)
+		)
+
+		if(location=="" and email==""):
+			books = Book.query.filter_by(category=category)
+		elif(category=="" and email==""):
+			books = Book.query.filter_by(location=location)
+		elif(category=="" and location==""):
+			books = Book.query.filter_by(email=email)
+		elif(email==""):
+			books = Book.query.filter_by(location=location, category = category)
+		elif(location==""):
+			books = Book.query.filter_by(email=email, category = category)
+		elif(category==""):
+			books = Book.query.filter_by(location=location, email = email)
+		else:
+			books = Book.query.filter_by(location=location, email=email, category = category)
 	return render_template('index.html', books = books, locations = LOCATIONS, categories = CATEGORIES)
 
 @app.route('/donate-book')
-def donate_bool():
+def donate_book():
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/donate-book", datetime.now())
+	)
+
 	return render_template('donate_book.html', locations = LOCATIONS, categories = CATEGORIES)
 
 @app.route('/about-us')
 def about_us():
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/about-us", datetime.now())
+	)
+
 	return render_template('about_us.html', locations = LOCATIONS, categories = CATEGORIES)
 
 @app.route('/donation-form', methods=['POST'])
 def donation_form():
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/donation-form", datetime.now())
+	)
+	
 	username = request.form['form-name']
 	email = request.form['form-email']
 	phone = request.form['form-phone']
@@ -83,7 +127,6 @@ def donation_form():
 	category = request.form['form-category']
 	description = request.form['form-description']
 	location = request.form['form-location']
-	address = request.form['form-address']
 	try:
 		file = request.files['file']
 		if(file.filename==""):
@@ -102,8 +145,7 @@ def donation_form():
 			title = title, 
 			category = category, 
 			description = description, 
-			phone = phone, 
-			address = address, 
+			phone = phone,
 			location = location, 
 			image_path = filename
 		)
@@ -118,41 +160,70 @@ def donation_form():
 	return redirect('/')
 
 
+@app.route('/add-comment', methods=['POST'])
+def add_comment():
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/add-comment", datetime.now())
+	)
+
+	comment = request.form['form-comment']
+	id = request.form['form-id']
+
+	try:
+		comment = Comments(
+			pid = id,
+			comment = comment
+		)
+
+		db.session.add(comment)
+		db.session.commit()
+		flash("Comment posted successfully!","alert-success")
+	except:
+		flash("Please try again later.","alert-warning")
+
+	return redirect('/book/{}/get'.format(id))
+
 @app.route('/book/<id>/get')
 def show_book(id):
-	return render_template('show_book.html', book = Book.query.filter_by(id = id)[0], locations = LOCATIONS, categories = CATEGORIES)
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/book/{}/get".format(id), datetime.now())
+	)
+
+	book = Book.query.filter_by(id = id)
+	print(type(book))
+	# return "Abc"
+	return render_template('show_book.html', book = book[0], locations = LOCATIONS, categories = CATEGORIES, comments = Comments.query.filter_by(pid = id))
 
 @app.route('/book/<id>/request', methods=['POST'])
 def request_book(id):
 	book = Book.query.filter_by(id = id)[0]
 	username = request.form['form-name']
-	email = request.form['form-email']
 	phone = request.form['form-phone']
-	school = request.form['form-school']
 	reason = request.form['form-reason']
-	location = request.form['form-location']
 	address = request.form['form-address']
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/requested_books.csv",
+		(username, phone, reason, address)
+	)
+
 	message = '''
 Hi {},\n
 Request for your book {} on Share Kitaab.\n
 Details:\n
 	Requested by: {}
-	Email ID: {}
 	Phone Number: {}
-	School Name: {}
-	Reason: {}	
-	Loaction: {}
-	Address: {}
+	Reason: {}
 	'''.format(
 		book.username,
 		book.title,
 		username,
-		email,
 		phone,
-		school,
 		reason,
-		location,
-		address
 	)
 	try:
 		s.sendmail(me, book.email, message)	
@@ -163,6 +234,12 @@ Details:\n
 
 @app.route('/book/delete/verify', methods=['POST'])
 def delete_book_verify():
+
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/book/delete/verify", datetime.now())
+	)
+	
 	id = request.form['form-id']
 	book = Book.query.filter_by(id = id)[0]
 	email = request.form['form-email']
@@ -171,18 +248,48 @@ def delete_book_verify():
 			image_path = os.path.join(app.config['UPLOAD_FOLDER'], book.image_path)
 			db.session.delete(book)
 			db.session.commit()
-			os.remove(image_path)
+
+			csv_writer.write(
+				"/home/raghav/Desktop/kitab_app/user_data/deleted_cards.csv",
+				book.get_book_as_tuple("verified")
+			)
+
+			if(os.path.exists(image_path) and os.path.isfile(image_path)):
+				os.remove(image_path)
+
 			flash("Deleted successfully!!","alert-success")
 		except:
+
+			csv_writer.write(
+				"/home/raghav/Desktop/kitab_app/user_data/failed_delete.csv",
+				(id, email, datetime.now())
+			)
+
 			flash("Please try again later!!","alert-danger")
 	else:
+
+		csv_writer.write(
+			"/home/raghav/Desktop/kitab_app/user_data/failed_delete.csv",
+			(id, email, datetime.now())
+		)
+		
 		flash("Please enter your registered Eamil ID!!","alert-warning")
 
 	return redirect('/')
 
 @app.route('/book/<id>/delete')
 def delete_book(id):
+	
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/url_hits.csv",
+		("/book/{}/delete".format(id), datetime.now())
+	)
+
 	book = Book.query.filter_by(id = id)[0]
+	csv_writer.write(
+		"/home/raghav/Desktop/kitab_app/user_data/deleted_cards.csv",
+		book.get_book_as_tuple("not verified")
+	)
 	image_path = os.path.join(app.config['UPLOAD_FOLDER'], book.image_path)
 	db.session.delete(book)
 	db.session.commit()
